Sửa bug lịch

+++ EasyTimeTable/Views/Student/Tuition/CustomYesNoDialog.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using EasyTimeTable.ViewModel;
+using Syncfusion.Windows.Controls.Input;
 using System;
 using System.Collections.Generic;
 using System.Configuration;
@@ -24,37 +25,61 @@ namespace EasyTimeTable.Views.Student.Tuition
     {
         private string message;
         private string name;
+        private bool huy;
 
         public CustomYesNoDialog()
         {
             InitializeComponent();
         }
 
-        public CustomYesNoDialog(string message, string name)
+        public CustomYesNoDialog(string message, string name, bool huy)
         {
             InitializeComponent();
-            txt.Text = "Bạn có chắc hủy môn " + message + " - " + name;
-            this.message = message;
-            this.name = name;
+            this.huy = huy;
+            if (huy)
+            {
+                txt.Text = "Bạn có chắc hủy môn " + message + " - " + name;
+                this.message = message;
+                this.name = name;
+            }
+            else
+            {
+                txt.Text = "Bạn có chắc chắn muốn thanh toán học phí (bạn không thể đăng kí thêm môn học)";
+                Title.Text = "Xác nhận thanh toán";
+            }
         }
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            SqlConnection con = new SqlConnection(ConfigurationManager.ConnectionStrings["ConnectionString"].ConnectionString);
-            con.Open();
-            var cmd = new SqlCommand("delete from lophocphansinhvien where mahocphan like @mahocphan and masv = @masv", con);
-            cmd.Parameters.Add("@mahocphan", System.Data.SqlDbType.VarChar);
-            cmd.Parameters["@mahocphan"].Value = message + "%";
-            cmd.Parameters.Add("@masv", System.Data.SqlDbType.VarChar);
-            cmd.Parameters["@masv"].Value = LoginViewModel.mssv;
-            var dr = cmd.ExecuteNonQuery();
-            MessageBox.Show("Đã hủy đăng kí môn: " + message + " - " + name);
             var viewModel = (TuitionViewModel)DataContext;
-            if (viewModel.LoadDB.CanExecute(null))
-                viewModel.LoadDB.Execute(null);
-            if (viewModel.LoadListDB.CanExecute(null))
-                viewModel.LoadListDB.Execute(null);
-            this.Close();
+            if (huy)
+            {
+                SqlConnection con = new SqlConnection(ConfigurationManager.ConnectionStrings["ConnectionString"].ConnectionString);
+                con.Open();
+                var cmd = new SqlCommand("delete from lophocphansinhvien where mahocphan like @mahocphan and masv = @masv", con);
+                cmd.Parameters.Add("@mahocphan", System.Data.SqlDbType.VarChar);
+                cmd.Parameters["@mahocphan"].Value = message + "%";
+                cmd.Parameters.Add("@masv", System.Data.SqlDbType.VarChar);
+                cmd.Parameters["@masv"].Value = LoginViewModel.mssv;
+                var dr = cmd.ExecuteNonQuery();
+                MessageBox.Show("Đã hủy đăng kí môn: " + message + " - " + name);
+                if (viewModel.LoadDB.CanExecute(null))
+                    viewModel.LoadDB.Execute(null);
+                if (viewModel.LoadListDB.CanExecute(null))
+                    viewModel.LoadListDB.Execute(null);
+                this.Close();
+            }
+            else
+            {
+                SqlConnection con = new SqlConnection(ConfigurationManager.ConnectionStrings["ConnectionString"].ConnectionString);
+                con.Open();
+                var cmd = new SqlCommand("update lophocphansinhvien set ngaythanhtoan = @ngaythanhtoan where masv = '" + LoginViewModel.mssv + "' and ngaythanhtoan is null and daduyet = 0", con);
+                cmd.Parameters.Add("@ngaythanhtoan", System.Data.SqlDbType.DateTime);
+                cmd.Parameters["@ngaythanhtoan"].Value = DateTime.Now;
+                cmd.ExecuteNonQuery();
+                MessageBox.Show("Đã thanh toán thành công");
+                this.Close();
+            }
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
